+++ src/feintrack_d.h
@@ -2,7 +2,7 @@
 
 ////////////////////////////////////////////////////////////////////////////////
 
-typedef long mask_type;
+typedef int32_t mask_type;
 ////////////////////////////////////////////////////////////////////////////////
 
 // Выровненная структура, соответствующая цвету пикселя
@@ -61,7 +61,7 @@ typedef struct
 ////////////////////////////////////////////////////////////////////////////////
 
 // Вычитание фона без обновления заднего плана
-void back_substraction(long *bgr32, float *params_b_mu, float *params_b_sigma, float *params_g_mu, float *params_g_sigma, float *params_r_mu, float *params_r_sigma,
+void back_substraction(int32_t *bgr32, float *params_b_mu, float *params_b_sigma, float *params_g_mu, float *params_g_sigma, float *params_r_mu, float *params_r_sigma,
 					   mask_type *mask, int frame_width, int frame_height, float eps_b, float eps_g, float eps_r);
 
 void back_substraction_gray(float *p_val, float *params_mu, float *params_sigma, mask_type *mask, int frame_width, int frame_height, float eps);
@@ -69,27 +69,27 @@ void back_substraction_gray(float *p_val, float *params_mu, float *params_sigma,
 ////////////////////////////////////////////////////////////////////////////////
 
 // Вычитание фона с обновлением заднего плана
-void back_substraction_upd(long *bgr32, float *params_b_mu, float *params_b_sigma, float *params_g_mu, float *params_g_sigma, float *params_r_mu, float *params_r_sigma,
+void back_substraction_upd(int32_t *bgr32, float *params_b_mu, float *params_b_sigma, float *params_g_mu, float *params_g_sigma, float *params_r_mu, float *params_r_sigma,
 						   mask_type *mask, int frame_width, int frame_height, float eps_b, float eps_g, float eps_r, PixelUpdateParams pup);
 
 void back_substraction_gray_upd(float *p_val, float *params_mu, float *params_sigma, mask_type *mask, int frame_width, int frame_height, float eps, PixelUpdateParams pup);
 ////////////////////////////////////////////////////////////////////////////////
 
 // Вычитание фона с помощью смеси гауссианов
-void back_substraction_mixture(long* bgr32, BgrndProcess* process1, BgrndProcess* process2, BgrndProcess* process3,
-							   long* curr_processes, long* created_processes, mask_type* mask,
+void back_substraction_mixture(int32_t* bgr32, BgrndProcess* process1, BgrndProcess* process2, BgrndProcess* process3,
+                               int32_t* curr_processes, int32_t* created_processes, mask_type* mask,
 							   int frame_width, int frame_height, float eps_b, float eps_g, float eps_r,
 							   PixelUpdateParams pup, MixturePixelUpdateParams mup);
 
 // Вычитание фона с помощью смеси гауссианов
 void back_substraction_mixture_gray(float* p_val, BgrndProcess* process1, BgrndProcess* process2, BgrndProcess* process3,
-									long* curr_processes, long* created_processes, mask_type* mask,
+                                    int32_t* curr_processes, int32_t* created_processes, mask_type* mask,
 									int frame_width, int frame_height, float eps,
 									PixelUpdateParams pup, MixturePixelUpdateParams mup);
 ////////////////////////////////////////////////////////////////////////////////
 
 // Сброс статистики в указанной области
-void reset_statistic(long *bgr32, float *params_b_mu, float *params_b_sigma, float *params_g_mu, float *params_g_sigma, float *params_r_mu, float *params_r_sigma, int frame_width,
+void reset_statistic(int32_t *bgr32, float *params_b_mu, float *params_b_sigma, float *params_g_mu, float *params_g_sigma, float *params_r_mu, float *params_r_sigma, int frame_width,
 					 int left, int right, int top, int bottom, float max_sigma_val);
 
 void reset_statistic_gray(float *p_val, float *params_mu, float *params_sigma, int frame_width,
@@ -97,7 +97,7 @@ void reset_statistic_gray(float *p_val, float *params_mu, float *params_sigma, i
 ////////////////////////////////////////////////////////////////////////////////
 
 // Обновление статистики в указанной области
-void update_statistic(long *bgr32, float *params_b_mu, float *params_b_sigma, float *params_g_mu, float *params_g_sigma, float *params_r_mu, float *params_r_sigma, int frame_width,
+void update_statistic(int32_t *bgr32, float *params_b_mu, float *params_b_sigma, float *params_g_mu, float *params_g_sigma, float *params_r_mu, float *params_r_sigma, int frame_width,
 					  int left, int right, int top, int bottom, PixelUpdateParams pup);
 
 void update_statistic_gray(float *p_val, float *params_mu, float *params_sigma, int frame_width,
